+++ R/listTumorFractions.R
@@ -1,6 +1,6 @@
 #' List of available tumor fractions for a data set
 #'
-#' @param dataSet  The name of a data set from the package, see
+#' @param dataSet  The name of a data set from the package, see \link{listDataSets}
 #'
 #' @return A numeric vector, the available tumor fractions for a data set
 #' @export
@@ -10,17 +10,14 @@
 #' fracs <- listTumorFractions(dataSets[1])
 #'
 listTumorFractions <- function(dataSet) {
-    ds <- listDataSets()
-    if( !(dataSet %in% ds)) {
-        stop("Argument 'dataSet' should be one of ", paste(ds, collapse=","))
+    dataSets <- listDataSets()
+    if( !(dataSet %in% dataSets)) {
+        stop("Argument 'dataSet' should be one of ", paste(dataSets, collapse=", "))
     }
     path <- system.file("extdata", package="acnr")
-    patt <- "(.*),([0-9]+),cnRegions.rds"
-
-    dpath <- file.path(path, dataSet)
-    chipType <- list.files(dpath)
-    dpath <- file.path(dpath, chipType)
-    fl <- list.files(dpath, pattern=patt)
-    pcts <- as.numeric(gsub(patt, "\\2", fl))/100
-    sort(pcts)
+    filename <- sprintf("%s.rds", dataSet)
+    pathname <- file.path(path, filename)
+    dat <- readRDS(pathname)
+    pcts <- unique(dat$cellularity)
+    sort(pcts, decreasing=TRUE)
 }