+++ inst/preprocessing/GSE11976/02.exportRegionData.R
@@ -1,39 +1,38 @@
-library(R.utils)
-source("01.defineCopyNumberSegments.R")
+sf <- system.file("preprocessing/GSE11976/01.defineCopyNumberSegments.R", package="acnr")
 str(regDat)
 
 datI <- readRDS("CRL2324_dilutionSeries.rds")
-datI$posMb <- datI$Position/1e6;
-## Define genotypes
+datI$posMb <- datI$position/1e6;
+nms <- names(datI)[grep("^LRR", names(datI))]
+pcts <- gsub("LRR_([0-9]+)", "\\1", nms)
 
-
-pct <- c("100","79", "50", "34", "0") 
-dataSet <- "CRL2324,BAF"
+sampleName <- "CRL2324"
+dataSet <- "GSE11976,BAF"
 chipType <- "HumanCNV370v1"
 savPath <- "cnRegionData"
 savPath <- Arguments$getWritablePath(savPath)
 path <- file.path(savPath, dataSet, chipType);
 path <- Arguments$getWritablePath(path)
-for(pp in pct){
-    sampleName <- sprintf("CRL2324,BAF,%s",pp)
-    for (rr in 1:nrow(regDat)) {
-        chr <- regDat[rr, "chr"]
-        reg <- c(regDat[rr, "begin"], regDat[rr, "end"])
-        type <- regDat[rr, "type"]
-        datCC <- subset(datI, chromosome==chr)
-        datRR <- subset(datCC, posMb>=reg[1] & posMb<=reg[2])
-        opos <- order(datRR$posMb)
-        datRRo <- datRR[opos, ]
-        datRRo <- na.omit(datRRo)
+
+for (rr in 1:nrow(regDat)) {
+    chr <- regDat[rr, "chr"]
+    reg <- c(regDat[rr, "begin"], regDat[rr, "end"])
+    type <- regDat[rr, "type"]
+    datCC <- subset(datI, chromosome==chr)
+    datRR <- subset(datCC, posMb>=reg[1] & posMb<=reg[2])
+    opos <- order(datRR$posMb)
+    datRRo <- datRR[opos, ]
+    datRRo <- na.omit(datRRo)
         
-        columns <- c(sprintf("logR_%s",pp), sprintf("baf_%s",pp), "muN")
-        datSS <- na.omit(datRRo[, columns])
+    for(pp in pcts){
+        cols <- c(sprintf("LRR_%s", pp), sprintf("BAF_%s", pp), "muN")
+        datSS <- na.omit(datRRo[, cols])
         names(datSS) <- c("c","b","genotype")
         datSS$c <- 2^(datSS$c + 1)
         tgb <- tan(datSS$b*pi/2)
         datSS$b <- tgb/(1+tgb)
         datSS <- round(datSS, 3)
-        filename <- sprintf("%s,%s.rds", sampleName, type)
+        filename <- sprintf("%s,%s,%s.rds", sampleName, pp, type)
         pathname <- file.path(path,filename)
         saveRDS(datSS, file=pathname)
     }