+++ inst/preprocessing/GSE13372/03.doASCRMAv2.R
@@ -17,11 +17,18 @@ print(cdf)
 csR <- AffymetrixCelSet$byName(dataSet, chipType=chipType, cdf=cdf);
 print(csR);
 
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 # AS-CRMAv2
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-nn <- names(csR)
-resR <- doASCRMAv2(csR[which(nn=="GSM337641")], verbose=verbose);
-resN <- doASCRMAv2(csR[which(nn=="GSM337663")], verbose=verbose);
-resRef <- doASCRMAv2(csR[which(nn=="GSM337662")], verbose=verbose);
+
+## Extract only one tumor and two matching normals
+nms <- c(T="GSM337641", N="GSM337662", R="GSM337663")
+idxs <- indexOf(csR, nms)
+csE <- extract(csR, idxs)
+
+res <- doASCRMAv2(csE, verbose=verbose)
+
+# nn <- names(csR)
+# resR <- doASCRMAv2(csR[which(nn=="GSM337641")], verbose=verbose);
+# resN <- doASCRMAv2(csR[which(nn=="GSM337663")], verbose=verbose);
+# resRef <- doASCRMAv2(csR[which(nn=="GSM337662")], verbose=verbose);