+++ inst/preprocessing/GSE13372/04.writeWholeGenomeData.R
@@ -3,38 +3,29 @@
 dataSet <- "GSE13372";
 chipType <- "GenomeWideSNP_6"
 
-if (FALSE) {
-    rpn <- sprintf("testScripts/system/preprocessing/%s/03.doASCRMAv2.R", dataSet)
-    pn <- system.file(rpn, package="acnr")
-    file.exists(pn)
-    source(pn)
-}
+tags <- "ACC,ra,-XY,BPN,-XY,AVG,FLN,-XY"
+ds <- AromaUnitTotalCnBinarySet$byName(dataSet, tags=tags, chipType=chipType)
+length(ds)
 
-## tumor samples
-dsT <- resR$total
-dsN <- resN$total
-dsRef <- resRef$total
-
+dsE <- extract(ds, indexOf(ds, nms))
 tumorName <- "HCC1143_GLEYS"
 normalName <- "HCC1143BL_GLEYS"
+sampleName <- "HCC1143"
 
 ## Naming
 ## Extract (total,beta) estimates for the tumor-normal pair
-dataT <- extractPSCNArray(dsT);
-str(dataT);
-dataN <- extractPSCNArray(dsN);
-str(dataN);
-dataRef <- extractPSCNArray(dsRef);
-str(dataRef);
+dat <- extractPSCNArray(dsE);
+dn3 <- dimnames(dat)[[3]]
+dimnames(dat)[[3]] <- names(nms)[match(dn3, nms)]
 
 ## Get (chromosome, position) annotation data
-ugp <- getAromaUgpFile(dsT);
+ugp <- getAromaUgpFile(dsE);
 chromosome <- ugp[,1,drop=TRUE];
 x <- ugp[,2,drop=TRUE];
 
 ## Total intensities and Allele B fractions for the normal
-thetaN <- dataRef[,"total", 1];
-betaN <- dataRef[,"fracB", ];
+thetaN <- dat[,"total", "R"];
+betaN <- dat[,"fracB", "R" ];
 
 datPath <- "wholeGenomeData";
 ## A symbolic link to "/home/share/Data/wholeGenomeData"
@@ -46,15 +37,18 @@ path <- file.path(datPath, dataSet, chipType);
 path <- Arguments$getWritablePath(path);
 
 
-idxs <- seq(length=dim(dataT)[3])
+idxs <- c("N", "T")
 for (ss in idxs) {
-    pairName <- sprintf("%svs%s,100", tumorName, normalName)
-
+    if (ss == "N") {
+        pairName <- sprintf("%svs%s,0", tumorName, normalName)
+    } else if (ss=="T") {
+        pairName <- sprintf("%svs%s,100", tumorName, normalName)
+    }
     ## Total CNs for the tumor relative to the matched normal
-    CT <- 2 * dataT[,"total", ss] / thetaN;
+    CT <- 2 * dat[,"total", ss] / thetaN;
     
     ## Allele B fractions for the tumor
-    betaT <- dataT[,"fracB", ss];
+    betaT <- dat[,"fracB", ss];
     
     ## Setup data structure
     df <- data.frame(chromosome=chromosome, x=x, CT=CT, betaT=betaT, betaN=betaN);
@@ -68,53 +62,19 @@ for (ss in idxs) {
     saveRDS(dfC, file=pathname)
     
     ## chromosome by chromosome
-    for (cc in 1:24) {
-        print(cc)
-        fileName <- sprintf("%s,chr=%02d.rds", pairName, cc)
-        pathname <- file.path(path, fileName)
-        saveRDS(dfC, file=pathname)
-        
-        datCC <- subset(dfC, chromosome==cc)
-        o <- order(datCC$x)
-        datCC <- datCC[o, ]
-        str(datCC)
-        save(datCC, file=pathname)
+    if (FALSE) {
+        for (cc in 1:24) {
+            print(cc)
+            fileName <- sprintf("%s,chr=%02d.rds", pairName, cc)
+            pathname <- file.path(path, fileName)
+            saveRDS(dfC, file=pathname)
+            
+            datCC <- subset(dfC, chromosome==cc)
+            o <- order(datCC$x)
+            datCC <- datCC[o, ]
+            str(datCC)
+            save(datCC, file=pathname)
+        }
     }
 }
 
-
-idxs <- seq(length=dim(dataN)[3])
-for (ss in idxs) {
-    pairName <- sprintf("%svs%s,0", tumorName, normalName)
-
-    ## Total CNs for the tumor relative to the matched normal
-    CT <- 2 * dataN[,"total", ss] / thetaN;
-    
-    ## Allele B fractions for the tumor
-    betaT <- dataN[,"fracB", ss];
-    
-    ## Setup data structure
-    df <- data.frame(chromosome=chromosome, x=x, CT=CT, betaT=betaT, betaN=betaN);
-
-    dfC <- subset(df, !is.na(df$chromosome) & !is.na(df$x))
-    str(dfC)
-    
-    ## save  
-    fileName <- sprintf("%s.rds", pairName)
-    pathname <- file.path(path, fileName)
-    saveRDS(dfC, file=pathname)
-    
-    ## chromosome by chromosome
-    for (cc in 1:24) {
-        print(cc)
-        fileName <- sprintf("%s,chr=%02d.rds", pairName, cc)
-        pathname <- file.path(path, fileName)
-        saveRDS(dfC, file=pathname)
-        
-        datCC <- subset(dfC, chromosome==cc)
-        o <- order(datCC$x)
-        datCC <- datCC[o, ]
-        str(datCC)
-        save(datCC, file=pathname)
-    }
-}