+++ inst/preprocessing/GSE29172/08.mergeCnRegionData.R
@@ -3,14 +3,12 @@ library("R.utils");
 dataSet <- "GSE29172"
 chipType <- "GenomeWideSNP_6"
 
-if (FALSE) {
-    ## Define CN regions:
-    rpn <- sprintf("preprocessing/%s/05.defineCopyNumberSegments.R", dataSet)
-    pn <- system.file(rpn, package="acnr")
-    file.exists(pn)
-    source(pn)
-    str(regDat)
-}
+## Define CN regions
+regFile <- "05.defineCopyNumberSegments.R"
+pn <- file.path("preprocessing", dataSet, regFile)
+sf <- system.file(pn, package="acnr")
+source(sf)
+str(regDat)
 
 regPath <- "cnRegionData";
 regPath <- Arguments$getReadablePath(regPath);
@@ -22,34 +20,37 @@ path <- file.path(regPath, ds, chipType);
 path <- Arguments$getReadablePath(path);
 
 sampleName <- "H1395vsBL1395"
+
 pattern <- sprintf("%s,([0-9]+),\\(([0-9]),([0-9])\\).rds", sampleName)
 filenames <- list.files(path, pattern=pattern)
 pcts <- unique(gsub(pattern, "\\1", filenames))
 
-savPath <- file.path("extdata", "GSE29172", chipType)
-savPath <- Arguments$getWritablePath(savPath);
+savPath <- Arguments$getWritablePath("inst/extdata")
 
 types <- regDat[["type"]]
+datList <- list()
 for (pct in pcts) {
-  print(pct)
-  pattern <- sprintf("H1395vsBL1395,%s,(.*).rds", pct)
-  filenames <- list.files(path, pattern=pattern)
-  types <- gsub(pattern, "\\1", filenames)
-
-  pathnames <- file.path(path, filenames)
-  names(pathnames) <- types
-  datList <- lapply(pathnames, readRDS)
-  dat <- NULL
-  for (dd in names(datList)) {
-    datDD <- datList[[dd]]
-    datDD$region <- dd
-    dat <- rbind(dat, datDD)
-  }
-  str(dat)
-  rownames(dat) <- NULL
-  filename <- sprintf("%s,%s,cnRegions.rds", sampleName, pct)
-  pathname <- file.path(savPath, filename)
-  saveRDS(dat, file=pathname)
+    print(pct)
+    pattern <- sprintf("%s,%s,(.*).rds", sampleName, pct)
+    filenames <- list.files(path, pattern=pattern)
+    types <- gsub(pattern, "\\1", filenames)
+    
+    pathnames <- file.path(path, filenames)
+    for (tt in seq(along=types)) {
+        pathname <- pathnames[tt]
+        typ <- types[tt]
+        dat <- readRDS(pathname)
+        dat$region <- typ
+        dat$cellularity <- as.numeric(pct)/100
+        tag <- sprintf("%s,%s", pct, typ)
+        datList[[tag]] <- dat
+    } 
 }
+dat <- do.call("rbind", datList)
+rownames(dat) <- NULL
+str(dat)
 
-
+dsName <- "GSE29172_H1395"
+filename <- sprintf("%s.rds", dsName)
+pathname <- file.path(savPath, filename)
+saveRDS(dat, file=pathname)