+++ NEWS.md
@@ -1,5 +1,12 @@
 # Package: acnr
 
+## Version: 0.3.0 [2016-12-15]
+
+* Data sets are now stored in `inst/extData`. 
+* Data sets are now documented individually.
+* Updated scripts for the preprocessing of data set GSE13372.
+* Now exporting 'betaT' and 'betaN' for data set GSE13372.
+
 ## Version: 0.2.8 [2016-11-25]
 
 * Added functions 'listDataSets' and 'listTumorFractions'